requirements.txt
# Core requirements for LlamaIndex knowledge base with hybrid search
llama-index==0.9.48
python-dotenv>=0.19.0
openai>=1.0.0
nltk>=3.6.0
python-magic>=0.4.24
tiktoken>=0.5.0
rank_bm25>=0.2.2

# Document format support (included by default)
# The KB will automatically detect available formats

# PDF support
#PyPDF2>=3.0.0
pypdf

# Excel support 
pandas>=2.0.0
openpyxl>=3.0.0

# Word document support
python-docx>=0.8.11
